requirements.txt
diffusers==0.32.2
transformers==4.46.3
accelerate==1.0.1
safetensors==0.5.3
pytorch-lightning==2.4.0
numpy==1.24.4
pandas==2.0.3
pillow==9.5.0
opencv-python==4.11.0.86
scikit-image==0.21.0
sentence-transformers
tokenizers==0.20.3
torchmetrics==1.5.2
scipy==1.10.1
scikit-learn==1.3.2
tqdm==4.67.1
ftfy==6.2.3
regex==2024.11.6
faiss-cpu==1.7.2
onnxruntime
sentencepiece
